date range fixes
