+++ platform/ui/src/components/DateRange/DateRange.css
@@ -21,23 +21,21 @@
 /** INPUT FIELD COMMON STYLES **/
 
 .DateInput_input {
-  /* background-image: url(../../assets/icons/calendar.svg); */
+  background-image: url('data:image/svg+xml;utf8,<svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24"><g fill="%236b6b6b" fill-rule="evenodd"><path d="M20 20h-4v-4h4v4zm-6-10h-4v4h4v-4zm6 0h-4v4h4v-4zm-12 6h-4v4h4v-4zm6 0h-4v4h4v-4zm-6-6h-4v4h4v-4zm16-8v22h-24v-22h3v1c0 1.103.897 2 2 2s2-.897 2-2v-1h10v1c0 1.103.897 2 2 2s2-.897 2-2v-1h3zm-2 6h-20v14h20v-14zm-2-7c0-.552-.447-1-1-1s-1 .448-1 1v2c0 .552.447 1 1 1s1-.448 1-1v-2zm-14 2c0 .552-.447 1-1 1s-1-.448-1-1v-2c0-.552.447-1 1-1s1 .448 1 1v2z"/></g></svg>');
+  background-size: 14px;
+  background-position: 10px center;
+  @apply bg-no-repeat;
 }
-
 .DateInput_input {
   @apply cursor-pointer border-custom-blue mt-2 bg-black shadow transition duration-300 appearance-none border-t border-l border-r border-b border-solid rounded w-full py-2 px-3 text-sm text-gray-700 leading-tight pl-8;
 }
-
 .DateInput_input:hover {
   @apply border-gray-500;
 }
-
 .DateInput_input:focus {
   @apply border-gray-500 outline-none;
 }
-
 /** FIRST INPUT STYLES **/
-
 .DateInput:first-child .DateInput_input {
   @apply rounded-r-none;
 }
@@ -48,14 +46,11 @@
 }
 
 /** SECOND INPUT STYLES **/
-
 .DateInput:last-child .DateInput_input {
   @apply rounded-l-none;
   margin-left: -1px;
 }
-
 /** ARROW STYLES **/
-
 .DateRangePickerInput_arrow {
   @apply hidden;
 }
@@ -68,25 +63,19 @@
 .PresetDateRangePicker_button {
   @apply m-0 py-2 px-3 bg-custom-blue border-0 rounded text-white text-base transition duration-300;
 }
-
 .PresetDateRangePicker_button:hover {
   @apply opacity-80;
 }
-
 /* SELECT MONTH PICKER */
 .DateRangePicker_select {
-  @apply bg-white text-custom-dark border rounded border-custom-gray appearance-none cursor-pointer text-base pl-2 pr-5 py-1;
-
-  /* NEEDED FOR ARROW DOWN */
+  @apply bg-white text-custom-dark border rounded border-custom-gray appearance-none cursor-pointer text-base pl-2 pr-5 py-1; /* NEEDED FOR ARROW DOWN */
   background-image: linear-gradient(45deg, transparent 50%, gray 50%),
     linear-gradient(135deg, gray 50%, transparent 50%);
   background-position: calc(100% - 11px) 11px, calc(100% - 6px) calc(11px);
   background-size: 5px 5px, 5px 5px;
   background-repeat: no-repeat;
 }
-
 /* CALENDAR DAYS */
-
 .CalendarDay {
   @apply border-0 rounded-full;
 }
@@ -97,20 +86,16 @@
 .CalendarDay__selected:hover {
   @apply bg-custom-blue text-white border-0 border-custom-blue;
 }
-
 .CalendarDay__blocked_out_of_range:hover {
   @apply border-0 bg-white cursor-not-allowed text-custom-gray;
 }
-
 /* MONTH NAVIGATION BUTTONS */
-
 .DayPickerNavigation_button__horizontalDefault,
 .DayPickerNavigation_button__horizontalDefault:hover {
   @apply border-custom-gray text-custom-gray;
   top: 24px;
   padding: 3px 9px;
 }
-
 .DayPickerNavigation_svg__horizontal {
   @apply fill-current;
 }
