+++ platform/ui/src/views/StudyList/components/StudyListFilter.js
@@ -1,7 +1,6 @@
 import React, { useState } from 'react';
 import PropTypes from 'prop-types';
 import classnames from 'classnames';
-import moment from 'moment';
 
 import { Button, Icon, Input, Typography, DateRange } from '@ohif/ui';
 
@@ -15,7 +14,10 @@ const defaultProps = {
   filtersValues: {
     patientName: '',
     mrn: '',
-    studyDate: '',
+    studyDate: {
+      startDate: null,
+      endDate: null,
+    },
     description: '',
     modality: '',
     accession: '',
@@ -80,7 +82,6 @@ const StudyListFilter = ({
     filtersValues
   );
   const { sortBy, sortDirection } = currentFiltersValues;
-  const [focusedInput, setFocusedInput] = useState(null);
 
   const handleFilterLabelClick = name => {
     let _sortDirection = 1;
@@ -128,74 +129,15 @@ const StudyListFilter = ({
   const renderInput = (inputType, name) => {
     switch (inputType) {
       case 'date-range': {
-        const today = moment();
-        const lastWeek = moment().subtract(7, 'day');
-        const lastMonth = moment().subtract(1, 'month');
-
-        const getDateEntry = (datePicked, rangeDatePicked) => {
-          return rangeDatePicked || datePicked || null;
-        };
-
-        const getDateEntryFromRange = (today, numOfDays, edge = 'start') => {
-          if (typeof numOfDays !== 'number') {
-            return;
-          }
-
-          if (edge === 'end') {
-            return today;
-          } else {
-            return today.subtract(numOfDays, 'days');
-          }
-        };
-
-        const studyListDateFilterNumDays = 7;
-
-        const defaultStartDate = getDateEntryFromRange(
-          today,
-          studyListDateFilterNumDays,
-          'start'
-        );
-        const defaultEndDate = getDateEntryFromRange(
-          today,
-          studyListDateFilterNumDays,
-          'end'
-        );
-
-        const studyDatePresets = [
-          {
-            text: 'Today',
-            start: today,
-            end: today,
-          },
-          {
-            text: 'Last 7 days',
-            start: lastWeek,
-            end: today,
-          },
-          {
-            text: 'Last 30 days',
-            start: lastMonth,
-            end: today,
-          },
-        ];
-
-        const studyDateFrom = {};
-        const studyDateTo = {};
-
         return (
           <div className="relative">
             <DateRange
-              startDate={getDateEntry(studyDateFrom, defaultStartDate)}
-              startDateId="start-date"
-              endDate={getDateEntry(studyDateTo, defaultEndDate)}
-              endDateId="end-date"
+              startDate={currentFiltersValues[name].startDate}
+              endDate={currentFiltersValues[name].endDate}
               onDatesChange={({ startDate, endDate, preset = false }) => {
                 console.log('studyDateFrom', startDate);
                 console.log('studyDateTo', endDate);
               }}
-              focusedInput={focusedInput}
-              onFocusChange={updatedVal => setFocusedInput(updatedVal)}
-              presets={studyDatePresets}
             />
           </div>
         );
@@ -318,15 +260,21 @@ const StudyListFilter = ({
 };
 
 StudyListFilter.propTypes = {
-  filtersMeta: PropTypes.arrayOf({
-    name: PropTypes.string,
-    dsplayName: PropTypes.string,
-    inputType: PropTypes.oneOf(['text', 'select', 'date-range', 'none']),
-    isSortable: PropTypes.bool,
-    gridCol: PropTypes.oneOf([1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12]),
-  }),
+  filtersMeta: PropTypes.arrayOf(
+    PropTypes.shape({
+      name: PropTypes.string,
+      dsplayName: PropTypes.string,
+      inputType: PropTypes.oneOf(['text', 'select', 'date-range', 'none']),
+      isSortable: PropTypes.bool,
+      gridCol: PropTypes.oneOf([1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12]),
+    })
+  ),
   filtersValues: PropTypes.object,
   numOfStudies: PropTypes.number,
+  studyDate: PropTypes.shape({
+    startDate: PropTypes.string,
+    endDate: PropTypes.string,
+  }),
 };
 
 export default StudyListFilter;
